requirements.txt
django==2.2.6
djangorestframework==3.10.3
django-rest-auth==0.9.5
django-allauth==0.40.0
pgcli==3.5.0
gunicorn==19.9.0
django-cors-headers==3.1.1
sentry-sdk==0.10.2
django-filter==2.2.0
django-oauth-toolkit==1.2.0
drf-tracking==1.5.0
djangorestframework-bulk==0.2.1
django-compat==1.0.15
xlwt==1.3.0
xmltodict==0.12.0
xlrd==2.0.1
cryptography==2.8
wheel==0.34.2
pandas
django-seed==0.3.1
psycopg2-binary==2.9.5